Refactor dbsync utils
